renamed DefaultPlatformChooser and moved it to the core, some small fixes
